+++ R/data.R
@@ -67,9 +67,9 @@
 "synapse_tables"
 
 
-#' Simulated fake synapse data for function examples and tests
+#' Simulated fake GENIE BPC data for function examples and tests
 #'
-#' A named list of simulated NSCLC clinical data
+#' A named list of simulated NSCLC clinical and genomic data
 #'
 #' @format A list of clinical data frames
 #' \describe{
@@ -80,6 +80,9 @@
 #'     \item{prissmm_imaging}{PRISSMM Imaging report data.frame}
 #'     \item{prissmm_pathology}{PRISSMM Pathology report data.frame}
 #'     \item{prissmm_md}{PRISSMM medical oncologist report data.frame}
-#'     \item{cpt}{CPT/NGS data.frame}
+#'     \item{cpt}{Cancer Panel Test (CPT)/Next Generation Sequencing (NGS) data.frame}
+#'     \item{mutations_extended}{Mutations data.frame}
+#'     \item{fusions}{Fusions data.frame}
+#'     \item{cpt}{Copy Number Alteration (CNA) data.frame}
 #' }
 "nsclc_test_data"